New apple image Removed Player.java
Reworked Food.java so that the foodObject spawns within the rectangles on the gameboard
Reworked Snake.java to work with new logic so the foodObjects spawn in the rectangles on the gameboard
Changed SnakeApplication.java so snake fits within the rectangles on gameboard.
Head is different colour. Added scoretext and highscoretext